DTO mismatch fix

+++ src/main/java/net/deepblacksea/train_api/api/dto/SessionDtos.java
@@ -1,4 +1,5 @@
 package net.deepblacksea.train_api.api.dto;
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.*;
 public class SessionDtos {
@@ -7,8 +8,8 @@ public class SessionDtos {
     public record SessionRes(UUID id, OffsetDateTime startedAt, String notes) {}
 
     //request payload for POST and GET /sessions/{id}/sets
-    public record SetCreateReq(UUID exerciseId, Integer reps, Double weightKg, Double rpe, String notes) {}
-    public record SetRes(UUID id, UUID exerciseId, Integer reps, Double weightKg, Double rpe, String notes) {}
+    public record SetCreateReq(UUID exerciseId, Integer reps, BigDecimal weightKg, BigDecimal rpe, String notes) {}
+    public record SetRes(UUID id, UUID exerciseId, Integer reps, BigDecimal weightKg, BigDecimal rpe, String notes) {}
 
     public record SessionWithSets(SessionRes session, List<SetRes> sets) {} //one session with all its sets
 }
